add fund formulas tests
