+++ tests/Feature/FundCriteriaTest.php
@@ -4,30 +4,31 @@ namespace Tests\Feature;
 
 use App\Models\Employee;
 use App\Models\Fund;
-use App\Models\FundCriterion;
 use App\Models\FundRequest;
 use App\Models\Identity;
-use App\Models\Organization;
-use App\Models\Prevalidation;
 use App\Models\RecordType;
+use App\Models\Voucher;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 use Illuminate\Foundation\Testing\WithFaker;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Arr;
 use Illuminate\Testing\TestResponse;
 use Tests\TestCase;
+use Tests\Traits\FundFormulaProductTestTrait;
+use Tests\Traits\MakesProviderAndProducts;
 use Tests\Traits\MakesTestFunds;
 use Tests\Traits\MakesTestIdentities;
 use Tests\Traits\MakesTestOrganizations;
 
 class FundCriteriaTest extends TestCase
 {
-    use WithFaker, DatabaseTransactions, MakesTestFunds, MakesTestIdentities, MakesTestOrganizations;
-
-    /**
-     * @var string
-     */
-    protected string $apiUrlCriteria = '/api/v1/platform/organizations/%s/funds/%s/criteria';
+    use WithFaker;
+    use MakesTestFunds;
+    use MakesTestIdentities;
+    use DatabaseTransactions;
+    use MakesTestOrganizations;
+    use MakesProviderAndProducts;
+    use FundFormulaProductTestTrait;
 
     /**
      * @var string
@@ -37,12 +38,12 @@ class FundCriteriaTest extends TestCase
     /**
      * @var string
      */
-    protected string $apiUrlPrevalidations = '/api/v1/platform/prevalidations';
+    protected string $apiUrlFundsApply = '/api/v1/platform/funds/%s/apply';
 
     /**
      * @var string
      */
-    protected string $apiUrlFundsApply = '/api/v1/platform/funds/%s/apply';
+    protected string $apiUrlFundsCheck = '/api/v1/platform/funds/%s/check';
 
     /**
      * @var string
@@ -297,13 +298,15 @@ class FundCriteriaTest extends TestCase
 
     /**
      * @return void
+     * @throws \Throwable
      */
     public function testFundRequestRecordsApplySuccess(): void
     {
+        $startDate = now();
         $organization = $this->makeTestOrganization($this->makeIdentity());
         $identity = $this->makeIdentity($this->makeUniqueEmail());
         $fund = $this->makeTestFund($organization);
-        $this->prepareTestFund($fund);
+        $this->prepareTestFundWithCriteria($fund);
 
         $response = $this->makeFundRequest($identity, $fund, [
             $this->makeRequestCriterionValue($fund, "test_bool", 'Ja'),
@@ -319,6 +322,10 @@ class FundCriteriaTest extends TestCase
 
         $response->assertSuccessful();
 
+        // create products for fund formula products assertion
+        // used 'test_number' as record key, created in method 'prepareTestFund'
+        $this->makeProviderAndProducts($fund, 'test_number');
+
         /** @var Employee $employee */
         $fundRequest = FundRequest::find($response->json('data.id'));
         $employee = $fundRequest->fund->organization->employees->first();
@@ -330,51 +337,103 @@ class FundCriteriaTest extends TestCase
         $fundRequest->approve($employee);
         $fundRequest->refresh();
 
-        $vouchers = $fundRequest->identity->vouchers;
-
         $this->assertTrue($fundRequest->isApproved());
-        $this->assertCount(1, $vouchers);
-        $this->assertEquals($vouchers[0]?->amount, $fund->fund_formulas[0]?->amount);
+        $vouchers = $fundRequest->identity->vouchers;
+        /** @var Voucher $budgetVoucher */
+        $budgetVoucher = $vouchers->whereNull('product_id')->first();
+        $this->assertNotNull($budgetVoucher);
+
+        $this->assertEquals(
+            $budgetVoucher->amount,
+            $fund->fund_formulas[0]?->amount + $fund->fund_formula_products->pluck('price')->sum()
+        );
+
+        $this->assertFundFormulaProducts($budgetVoucher, $startDate);
+}
+
+    /**
+     * @return void
+     * @throws \Throwable
+     */
+    public function testStoreSinglePrevalidationByRedeem()
+    {
+        $this->assertPrevalidationSuccess();
     }
 
     /**
      * @return void
+     * @throws \Throwable
      */
-    public function testStoreSinglePrevalidationSuccess(): void
+    public function testStoreSinglePrevalidationByApply()
     {
+        $this->assertPrevalidationSuccess(false);
+    }
+
+    /**
+     * @param bool $assertRedeem
+     * @return void
+     * @throws \Throwable
+     */
+    public function assertPrevalidationSuccess(bool $assertRedeem = true): void
+    {
+        $startDate = now();
         $organization = $this->makeTestOrganization($this->makeIdentity());
         $identity = $this->makeIdentity($this->makeUniqueEmail());
         $identity->setBsnRecord('123456789');
-        $identityHeaders = $this->makeApiHeaders($this->makeIdentityProxy($identity));
         $fund = $this->makeTestFund($organization);
-        $this->prepareTestFund($fund);
 
-        // create prevalidation
-        $response = $this->makeStorePrevalidationRequest($organization->identity, $fund, [
-            $this->makeRequestCriterionValue($fund, "test_bool", 'Ja'),
-            $this->makeRequestCriterionValue($fund, "test_iban", fake()->iban),
-            $this->makeRequestCriterionValue($fund, "test_date", '01-01-2010'),
-            $this->makeRequestCriterionValue($fund, "test_email", fake()->email),
-            $this->makeRequestCriterionValue($fund, "test_string", 'lorem_ipsum'),
-            $this->makeRequestCriterionValue($fund, "test_string_any", 'ipsum_lorem'),
-            $this->makeRequestCriterionValue($fund, "test_number", 7),
-            $this->makeRequestCriterionValue($fund, "test_select", 'foo'),
-            $this->makeRequestCriterionValue($fund, "test_select_number", 2),
-        ], [
-            'uid' => token_generator()->generate(32),
-        ]);
+        $identityHeaders = [
+            ...$this->makeApiHeaders($this->makeIdentityProxy($identity)),
+            'Client-Type' => 'webshop',
+            'Client-Key' => $fund->fund_config->implementation->key,
+        ];
 
+        $prevalidation = $this->buildFundAndPrevalidation($organization, $fund);
+
+        if ($assertRedeem) {
+            $code = $prevalidation->uid;
+            $response = $this->postJson($this->apiUrlFundsRedeem, compact('code'), $identityHeaders);
+            $response->assertSuccessful();
+            $voucher = $identity->vouchers()->whereNull('product_id')->first();
+        } else {
+            $prevalidation->assignToIdentity($identity);
+            $response = $this->postJson(sprintf($this->apiUrlFundsApply, $fund->id), [], $identityHeaders);
+            $response->assertSuccessful();
+            $voucher = Voucher::find($response->json('data.id'));
+        }
+
+        $this->assertNotNull($voucher);
+        $this->assertFundFormulaProducts($voucher, $startDate);
+    }
+
+    /**
+     * @return void
+     * @throws \Throwable
+     */
+    public function testPrevalidationCheckSuccess(): void
+    {
+        $bsn = '123456789';
+        $startDate = now();
+        $organization = $this->makeTestOrganization($this->makeIdentity(), ['bsn_enabled' => true]);
+        $identity = $this->makeIdentity($this->makeUniqueEmail());
+        $identity->setBsnRecord($bsn);
+        $fund = $this->makeTestFund($organization, [], ['csv_primary_key' => 'bsn']);
+
+        $identityHeaders = [
+            ...$this->makeApiHeaders($this->makeIdentityProxy($identity)),
+            'Client-Type' => 'webshop',
+            'Client-Key' => $fund->fund_config->implementation->key,
+        ];
+
+        $this->buildFundAndPrevalidation($organization, $fund, compact('bsn'));
+        $response = $this->postJson(sprintf($this->apiUrlFundsCheck, $fund->id), [], $identityHeaders);
         $response->assertSuccessful();
-        $prevalidation = Prevalidation::find($response->json('data.id'));
-        $code = $prevalidation->uid;
 
-        // redeem the prevalidation
-        $response = $this->postJson($this->apiUrlFundsRedeem, compact('code'), $identityHeaders);
-        $response->assertSuccessful();
+        /** @var Voucher $voucher */
+        $voucher = $identity->vouchers()->whereNull('product_id')->first();
 
-        // assert is possible to apply
-        $response = $this->postJson(sprintf($this->apiUrlFundsApply, $fund->id), [], $identityHeaders);
-        $response->assertSuccessful();
+        $this->assertNotNull($voucher);
+        $this->assertFundFormulaProducts($voucher, $startDate);
     }
 
     /**
@@ -385,7 +444,7 @@ class FundCriteriaTest extends TestCase
         $organization = $this->makeTestOrganization($this->makeIdentity());
         $identity = $this->makeIdentity($this->makeUniqueEmail());
         $fund = $this->makeTestFund($organization);
-        $this->prepareTestFund($fund);
+        $this->prepareTestFundWithCriteria($fund);
 
         $records = [
             $this->makeRequestCriterionValue($fund, "test_bool", 'Nee'),
@@ -412,7 +471,7 @@ class FundCriteriaTest extends TestCase
         $organization = $this->makeTestOrganization($this->makeIdentity());
         $identity = $this->makeIdentity($this->makeUniqueEmail());
         $fund = $this->makeTestFund($organization);
-        $this->prepareTestFund($fund);
+        $this->prepareTestFundWithCriteria($fund);
 
         $records = [
             $this->makeRequestCriterionValue($fund, "test_date", '01-01-1995'),
@@ -435,7 +494,7 @@ class FundCriteriaTest extends TestCase
         $organization = $this->makeTestOrganization($this->makeIdentity());
         $identity = $this->makeIdentity($this->makeUniqueEmail());
         $fund = $this->makeTestFund($organization);
-        $this->prepareTestFund($fund);
+        $this->prepareTestFundWithCriteria($fund);
 
         $records = [
             $this->makeRequestCriterionValue($fund, "test_date", '01-01-2030'),
@@ -458,7 +517,7 @@ class FundCriteriaTest extends TestCase
         $organization = $this->makeTestOrganization($this->makeIdentity());
         $identity = $this->makeIdentity($this->makeUniqueEmail());
         $fund = $this->makeTestFund($organization);
-        $this->prepareTestFund($fund);
+        $this->prepareTestFundWithCriteria($fund);
         $fund->criteria()->update([
             'optional' => true,
         ]);
@@ -487,7 +546,7 @@ class FundCriteriaTest extends TestCase
         $organization = $this->makeTestOrganization($this->makeIdentity());
         $identity = $this->makeIdentity($this->makeUniqueEmail());
         $fund = $this->makeTestFund($organization);
-        $this->prepareTestFund($fund);
+        $this->prepareTestFundWithCriteria($fund);
 
         $fund->criteria()->update([
             'show_attachment' => true,
@@ -506,127 +565,6 @@ class FundCriteriaTest extends TestCase
 
         $response = $this->makeFundRequest($identity, $fund, $records, false);
         $response->assertSuccessful();
-    }
-
-    /**
-     * @param Fund $fund
-     * @return void
-     */
-    protected function prepareTestFund(Fund $fund): void
-    {
-        $fund->criteria()->forceDelete();
-
-        $this->makeRecordType($fund->organization, RecordType::TYPE_BOOL, "test_bool");
-        $this->makeRecordType($fund->organization, RecordType::TYPE_IBAN, "test_iban");
-        $this->makeRecordType($fund->organization, RecordType::TYPE_DATE, "test_date");
-        $this->makeRecordType($fund->organization, RecordType::TYPE_EMAIL, "test_email");
-        $this->makeRecordType($fund->organization, RecordType::TYPE_STRING, "test_string");
-        $this->makeRecordType($fund->organization, RecordType::TYPE_STRING, "test_string_any");
-        $this->makeRecordType($fund->organization, RecordType::TYPE_NUMBER, "test_number");
-        $this->makeRecordType($fund->organization, RecordType::TYPE_SELECT, "test_select");
-        $this->makeRecordType($fund->organization, RecordType::TYPE_SELECT_NUMBER, "test_select_number");
-
-        $response = $this->updateCriteriaRequest([
-            $this->makeCriterion("test_bool", 'Ja', '='),
-            $this->makeCriterion("test_iban", null, '*'),
-            $this->makeCriterion("test_date", '01-01-2000', '>=', '01-01-1990', '01-01-2020'),
-            $this->makeCriterion("test_email", null, '*'),
-            $this->makeCriterion("test_string", 'lorem_ipsum', '=', 5, 20),
-            $this->makeCriterion("test_string_any", null, '*', 5, 20),
-            $this->makeCriterion("test_number", '7', '>=', 5, 10),
-            $this->makeCriterion("test_select", 'foo', '='),
-            $this->makeCriterion("test_select_number", 2, '>='),
-        ], $fund);
-
-        $fund->organization->forceFill([
-            'fund_request_resolve_policy' => $fund->organization::FUND_REQUEST_POLICY_AUTO_REQUESTED,
-        ])->save();
-
-        $response->assertSuccessful();
-    }
-
-    /**
-     * @param Organization $organization
-     * @param string $type
-     * @param string $key
-     * @return RecordType
-     */
-    protected function makeRecordType(
-        Organization $organization,
-        string $type,
-        string $key,
-    ): RecordType {
-         $recordType = RecordType::create([
-            'organization_id' => $organization->id,
-            'criteria' => true,
-            'type' => $type,
-            'key' => $key,
-        ]);
-
-         if ($type === $recordType::TYPE_SELECT) {
-             $recordType->record_type_options()->createMany([[
-                 'value' => 'foo',
-                 'name' => 'Foo',
-             ], [
-                 'value' => 'bar',
-                 'name' => 'Bar',
-             ]]);
-         }
-
-        if ($type === $recordType::TYPE_SELECT_NUMBER) {
-            $recordType->record_type_options()->createMany([[
-                'value' => 1,
-                'name' => 'Foo',
-            ], [
-                'value' => 2,
-                'name' => 'Bar',
-            ]]);
-        }
-
-         return $recordType;
-    }
-
-    /**
-     * @param string $key
-     * @param string|null $value
-     * @param string $operator
-     * @param string|null $min
-     * @param string|null $max
-     * @return array
-     */
-    protected function makeCriterion(
-        string $key,
-        ?string $value,
-        string $operator,
-        string $min = null,
-        string $max = null,
-    ): array {
-        return [
-            'max' => $max,
-            'min' => $min,
-            'value' => $value,
-            'operator' => $operator,
-            'record_type_key' => $key,
-            'show_attachment' => false,
-        ];
-    }
-
-    /**
-     * @param Fund $fund
-     * @param string $key
-     * @param string|null $value
-     * @return array
-     */
-    protected function makeRequestCriterionValue(Fund $fund, string $key, ?string $value): array
-    {
-        /** @var FundCriterion|null $criterion */
-        $criterion = $fund->criteria->firstWhere('record_type_key', $key);
-
-        return [
-            'fund_criterion_id' => $criterion?->id,
-            'value' => $value,
-            'files' => [],
-        ];
     }
 
     /**
@@ -721,48 +659,6 @@ class FundCriteriaTest extends TestCase
         $identity->setBsnRecord('123456789');
 
         return $this->postJson($url, compact('records'), $this->makeApiHeaders($proxy));
-    }
-
-    /**
-     * @param Identity $identity
-     * @param Fund $fund
-     * @param array $records
-     * @param array $extraData
-     * @return TestResponse
-     */
-    protected function makeStorePrevalidationRequest(
-        Identity $identity,
-        Fund $fund,
-        array $records,
-        array $extraData = [],
-    ): TestResponse {
-        $proxy = $this->makeIdentityProxy($identity);
-        $criteria = $fund->criteria()->pluck('record_type_key', 'id')->toArray();
-
-        return $this->postJson($this->apiUrlPrevalidations, [
-            'fund_id' => $fund->id,
-            'data' => [
-                ...array_reduce($records, fn ($list, $record) => [
-                    ...$list,
-                    $criteria[$record['fund_criterion_id']] => $record['value'],
-                    ], []),
-                ...$extraData,
-            ],
-        ], $this->makeApiHeaders($proxy));
-    }
-
-    /**
-     * @param array $criteria
-     * @param Fund $fund
-     * @return TestResponse
-     */
-    protected function updateCriteriaRequest(array $criteria, Fund $fund): TestResponse
-    {
-        $url = sprintf($this->apiUrlCriteria, $fund->organization_id, $fund->id);
-
-        return $this->patchJson($url, [
-            'criteria' => $criteria,
-        ], $this->makeApiHeaders($this->makeIdentityProxy($fund->organization?->identity)));
     }
 
     /**
