+++ tests/Traits/VoucherTestTrait.php
@@ -6,11 +6,7 @@ use App\Mail\Vouchers\VoucherAssignedBudgetMail;
 use App\Mail\Vouchers\VoucherAssignedProductMail;
 use App\Mail\Vouchers\VoucherAssignedSubsidyMail;
 use App\Models\Fund;
-use App\Models\FundCriterion;
-use App\Models\FundProvider;
 use App\Models\Identity;
-use App\Models\Prevalidation;
-use App\Models\Product;
 use App\Models\Voucher;
 use App\Models\VoucherTransaction;
 use App\Scopes\Builders\FundQuery;
@@ -20,13 +16,17 @@ use App\Traits\DoesTesting;
 use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Foundation\Testing\WithFaker;
-use Illuminate\Support\Collection as SupportCollection;
 use Illuminate\Support\Str;
 use Illuminate\Support\Arr;
 
 trait VoucherTestTrait
 {
-    use WithFaker, AssertsSentEmails, DoesTesting, MakesTestIdentities;
+    use WithFaker;
+    use DoesTesting;
+    use AssertsSentEmails;
+    use MakesTestIdentities;
+    use MakesProviderAndProducts;
+    use FundFormulaProductTestTrait;
 
     /**
      * @var string
@@ -39,47 +39,9 @@ trait VoucherTestTrait
     protected string $apiFundUrl = '/api/v1/platform/funds/%s';
 
     /**
-     * @var Product[]
-     */
-    protected array $approvedProducts = [];
-
-    /**
      * @var Identity[]
      */
     protected array $identities = [];
-
-    /**
-     * @var Product[]
-     */
-    protected array $unapprovedProducts = [];
-
-    /**
-     * @var Product[]
-     */
-    protected array $emptyStockProducts = [];
-
-    /**
-     * @param Fund $fund
-     * @param string|null $recordTypeKeyMultiplier
-     * @return void
-     */
-    protected function makeFundFormulaProducts(Fund $fund, ?string $recordTypeKeyMultiplier): void
-    {
-        if (!$fund->fund_formula_products->count()) {
-            array_map(function () use ($fund, $recordTypeKeyMultiplier) {
-                /** @var Product $product */
-                $product = array_random($this->approvedProducts);
-                $fund->fund_formula_products()->updateOrCreate([
-                    'product_id' => $product->id,
-                ], [
-                    'price' => $product->price,
-                    'record_type_key_multiplier' => $recordTypeKeyMultiplier,
-                ]);
-            }, range(0, 3));
-
-            $fund->load('fund_formula_products');
-        }
-    }
 
     /**
      * @param Fund $fund
@@ -110,9 +72,9 @@ trait VoucherTestTrait
                 $amount = $exceedVoucherAmountLimit ? $fund->getMaxAmountPerVoucher() + 10 : $amount;
             } elseif ($voucherType === 'product') {
                 $productId = match ($assert['product'] ?? 'approved') {
-                    'approved' => array_random($this->approvedProducts)->id,
-                    'unapproved' => array_random($this->unapprovedProducts)->id,
-                    'empty_stock' => array_random($this->emptyStockProducts)->id,
+                    'approved' => array_random($this->getApprovedProducts())->id,
+                    'unapproved' => array_random($this->getUnapprovedProducts())->id,
+                    'empty_stock' => array_random($this->getEmptyStockProducts())->id,
                     default => null,
                 };
             }
@@ -242,7 +204,7 @@ trait VoucherTestTrait
             $this->assertFieldsEquals($voucher, $voucherArr);
             $this->assertActivationCode($voucher, $voucherArr);
 
-            $this->assertFundFormulaProducts($voucher, $startDate, $assert['assign_by']);
+            $this->assertFundFormulaProducts($voucher, $startDate, $assert['assign_by'] === 'email');
         }
 
         if ($sameAssignBy > 0) {
@@ -292,32 +254,6 @@ trait VoucherTestTrait
         $voucherArr['activation_code']
             ? $this->assertNotNull($voucher->activation_code)
             : $this->assertNull($voucher->activation_code);
-    }
-
-    /**
-     * @param Voucher $voucher
-     * @param Carbon $startDate
-     * @param string $type
-     * @return void
-     */
-    protected function assertFundFormulaProducts(Voucher $voucher, Carbon $startDate, string $type): void
-    {
-        if ($voucher->isBudgetType()) {
-            foreach ($voucher->fund->fund_formula_products as $formulaProduct) {
-                $address = $voucher->identity_address;
-                $multiplier = $formulaProduct->getIdentityMultiplier($type === 'email' ? $address : null);
-
-                $productVoucherCount = Voucher::query()
-                    ->where('identity_address', $address)
-                    ->where('note', $voucher->note)
-                    ->where('product_id', $formulaProduct->product_id)
-                    ->where('created_at', '>=', $startDate)
-                    ->where('amount', $formulaProduct->price)
-                    ->count();
-
-                $this->assertEquals($multiplier, $productVoucherCount);
-            }
-        }
     }
 
     /**
@@ -531,97 +467,6 @@ trait VoucherTestTrait
     }
 
     /**
-     * @param Fund $fund
-     * @param string|null $recordTypeKeyMultiplier
-     * @return void
-     * @throws \Throwable
-     */
-    protected function makeProviderAndProducts(Fund $fund, ?string $recordTypeKeyMultiplier = null): void
-    {
-        $this->approvedProducts = $this->makeProducts($fund);
-        $this->emptyStockProducts = $this->makeProducts($fund, 0, 'global');
-        $this->unapprovedProducts = $this->makeProducts();
-
-        if ($fund->isTypeBudget()) {
-            $this->makeFundFormulaProducts($fund, $recordTypeKeyMultiplier);
-        }
-    }
-
-    /**
-     * @param Fund|null $fund
-     * @param int $stock
-     * @param string $allowProducts
-     * @return array
-     * @throws \Throwable
-     */
-    protected function makeProducts(
-        ?Fund $fund = null,
-        int $stock = 50,
-        string $allowProducts = 'individual',
-    ): array {
-        $testData = new TestData();
-        $identity = $this->makeIdentity($this->makeUniqueEmail('provider_'));
-        $provider = $testData->makeOrganizations("Provider", $identity->address)[0];
-
-        $products = $testData->makeProducts($provider, 5, [
-            'sold_out' => $stock === 0,
-            'total_amount' => $stock,
-            'unlimited_stock' => false,
-        ]);
-
-        $this->assertNotEmpty($products, 'Products not created');
-
-        if ($fund) {
-            /** @var FundProvider $fundProvider */
-            $fundProvider = $fund->providers()->firstOrCreate([
-                'state' => FundProvider::STATE_ACCEPTED,
-                'allow_budget' => $fund->isTypeBudget(),
-                'allow_products' => $allowProducts == 'global',
-                'organization_id' => $provider->id,
-            ]);
-
-            if ($allowProducts === 'individual') {
-                $this->updateProducts($fund, $fundProvider, [
-                    'enable_products' => array_map(fn (Product $product) => array_merge([
-                        'id' => $product->id,
-                        'limit_total' => rand(1, $stock),
-                        'limit_per_identity' => 1,
-                    ], $fund->isTypeSubsidy() ? [
-                        'amount' => $product->price,
-                    ] : []), $products),
-                ]);
-            }
-        }
-
-        return $products;
-    }
-
-    /**
-     * @param Fund $fund
-     * @param FundProvider $fundProvider
-     * @param array $params
-     * @return void
-     */
-    protected function updateProducts(
-        Fund $fund,
-        FundProvider $fundProvider,
-        array $params,
-    ): void {
-        $proxy = $this->makeIdentityProxy($fund->organization->identity);
-        $headers = $this->makeApiHeaders($proxy);
-
-        $url = sprintf(
-            $this->apiOrganizationUrl . '/funds/%s/providers/%s',
-            $fund->organization->id,
-            $fund->id,
-            $fundProvider->id
-        );
-
-        $response = $this->patch($url, $params, $headers);
-        $response->assertSuccessful();
-    }
-
-    /**
      * @param int $fundId
      * @return Fund
      */
@@ -641,65 +486,5 @@ trait VoucherTestTrait
     protected function randomFakeBsn(): int
     {
         return TestData::randomFakeBsn();
-    }
-
-    /**
-     * @param Identity $identity
-     * @param Identity $validator
-     * @return void
-     */
-    public function makeRecords(Identity $identity, Identity $validator): void
-    {
-        $prevalidations = Prevalidation::where([
-            'state' => 'pending',
-            'identity_address' => $identity->address
-        ])->get()->groupBy('fund_id')->map(static function(SupportCollection $arr) {
-            return $arr->first();
-        });
-
-        foreach ($prevalidations as $prevalidation) {
-            foreach ($prevalidation->prevalidation_records as $record) {
-                if ($record->record_type->key === 'bsn') {
-                    continue;
-                }
-
-                $identity
-                    ->makeRecord($record->record_type, $record->value)
-                    ->makeValidationRequest()
-                    ->approve($validator);
-            }
-
-            $prevalidation->update([
-                'state' => 'used'
-            ]);
-        }
-    }
-
-    /**
-     * @param Fund $fund
-     * @param Identity $identity
-     * @return void
-     * @throws \Throwable
-     */
-    public function makePrevalidation(Fund $fund, Identity $identity): void
-    {
-        $testData = new TestData();
-
-        if ($fund->fund_config->allow_prevalidations) {
-            $records = $fund->criteria->reduce(function (array $list, FundCriterion $criterion) {
-                return array_merge($list, [
-                    $criterion->record_type_key => match($criterion->operator) {
-                        '=' => intval($criterion->value),
-                        '>' => intval($criterion->value) + 1,
-                        '<' => intval($criterion->value) - 1,
-                        default => '',
-                    }
-                ]);
-            }, []);
-
-            $testData->makePrevalidations(
-                $identity, $fund, $testData->generatePrevalidationData($fund, 10, $records)
-            );
-        }
     }
 }
